Feat/Test/Fix: 고시원 매물 작성 추가
- 컬럼 수정에 따른 오류 수정

+++ src/main/java/com/kb/room/service/RoomTempService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface RoomTempService {
     List<RoomTempDTO> fetchAllRooms();
-    Long addRoom(GosiwonPostDTO dto); //고시원 작성
+    Integer addRoom(GosiwonPostDTO dto); //고시원 작성
 }
